+++ src/local.rs
@@ -7,6 +7,7 @@ mod object_id;
 mod position;
 mod room_coordinate;
 mod room_name;
+mod terrain;
 
 /// Represents two constants related to room names.
 ///
@@ -31,4 +32,5 @@ use crate::ROOM_SIZE;
 
 pub use self::{
     cost_matrix::*, lodash_filter::*, object_id::*, position::*, room_coordinate::*, room_name::*,
+    terrain::*,
 };